% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.snc.r
\name{print.snc}
\alias{print.snc}
\title{Print Method for SNC Objects}
\usage{
\method{print}{snc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"snc"} returned by the \code{\link{snc}} function.}

\item{...}{Ignored.}
}
\description{
Prints summary output for an object of class \code{"snc"}.
}
